requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
robotframework-seleniumlibrary==4.3.0  # via -r requirements.in (line 2)
robotframework==3.1.2     # via -r requirements.in (line 1), robotframework-seleniumlibrary
selenium==3.141.0         # via robotframework-seleniumlibrary
urllib3==1.25.8           # via selenium
